+++ src/Converter.php
@@ -19,6 +19,11 @@ class Converter
         $this->builder = $builder;
     }
 
+    /**
+     * @param CfdiData $cfdiData
+     * @param string $destination
+     * @return void
+     */
     public function createPdfAs(CfdiData $cfdiData, string $destination)
     {
         $this->builder->build($cfdiData, $destination);