+++ tests/PdfToText/PdfToText.php
@@ -35,6 +35,10 @@ class PdfToText
         return $shellExec->output();
     }
 
+    /**
+     * @param string $pdfFile
+     * @return string[]
+     */
     public function buildCommand(string $pdfFile): array
     {
         return [$this->pdftotext, '-eol', 'unix', '-raw', '-q', $pdfFile, '-'];